requirements.txt
mysql-connector-python
pyshark
openai
pandas
tabulate
python-dotenv
ipykernel
matplotlib
seaborn